requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
av==12.3.0
brotli
cachetools==5.5.2
certifi
charset-normalizer
contourpy==1.1.1
cycler==0.12.1
decord==0.6.0
deepspeed==0.17.5
einops==0.8.1
fonttools==4.57.0
ftfy==6.2.3
google-auth==2.40.3
google-auth-oauthlib==0.4.6
grpcio==1.70.0
hjson==3.1.0
idna
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
msgpack==1.1.1
ninja==1.13.0
nvidia-ml-py==13.580.65
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==1.5.3
pillow
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
PySocks
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.10.1
six==1.17.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==1.8
timm==0.4.12
torchsummary==1.5.1
tqdm==4.67.1
triton==1.0.0
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
